\name{alldata}
\alias{alldata}
\docType{data}
\title{
Data including base and expression data
}
\description{
Data including base and expression data
}
\usage{data("alldata")}
\format{
  A data frame with 100 observations on the following 200 variables.
  \describe{
   The first 100th variables are SNP and the second are expression data
   }
}
\examples{
data(alldata)
}
\keyword{datasets}
