\name{Tdata}
\alias{Tdata}
\docType{data}
\title{
Phenotype data
}
\description{
Phenotype data by rnorm function
}
\usage{data("Tdata")}
\format{
  A data frame with 100 observations on the following variable.
  \describe{
    \item{\code{Trait1}}{a numeric vector}
  }
}
\examples{
data(Tdata)
}
\keyword{datasets}